+++ Sources/SwiftSDK/Persistence/IDataStore.swift
@@ -25,11 +25,8 @@ protocol IDataStore {
     
     associatedtype CustomType
     
-    func save(entity: CustomType, responseHandler: ((CustomType) -> Void)!, errorHandler: ((Fault) -> Void)!)
     func save(entity: CustomType, isUpsert: Bool, responseHandler: ((CustomType) -> Void)!, errorHandler: ((Fault) -> Void)!)
-    func create(entity: CustomType, responseHandler: ((CustomType) -> Void)!, errorHandler: ((Fault) -> Void)!)
     func bulkCreate(entities: [CustomType], responseHandler: (([String]) -> Void)!, errorHandler: ((Fault) -> Void)!)
-    func update(entity: CustomType, responseHandler: ((CustomType) -> Void)!, errorHandler: ((Fault) -> Void)!)
     func bulkUpdate(whereClause: String?, changes: [String: Any], responseHandler: ((Int) -> Void)!, errorHandler: ((Fault) -> Void)!)
     func bulkUpsert(entities: [CustomType], responseHandler: (([String]) -> Void)!, errorHandler: ((Fault) -> Void)!)
     func removeById(objectId: String, responseHandler: ((Int) -> Void)!, errorHandler: ((Fault) -> Void)!)