+++ Tests/SwiftSDKTests/UOWTests/DataUOWBulkUpdateTests.swift
@@ -189,7 +189,7 @@ class DataUOWBulkUpdateTests: XCTestCase {
     func test_07_bulkUpdate() {
         let expectation = self.expectation(description: "PASSED: uow.bulkUpdate")
         let testObject = ["name": "Bob", "age": 25] as [String : Any]
-        Backendless.shared.data.ofTable("TestClass").save(entity: testObject, responseHandler: { createdObject in
+        Backendless.shared.data.ofTable("TestClass").save(entity: testObject, isUpsert: false, responseHandler: { createdObject in
             var objectToUpdate = createdObject
             objectToUpdate["age"] = 30
             let uow = UnitOfWork()