+++ Tests/SwiftSDKTests/UOWTests/DataUOWUpsertTests.swift
@@ -90,7 +90,7 @@ class DataUOWUpsertTests: XCTestCase {
     func test_04_upsert() {
         let expectation = self.expectation(description: "PASSED: uow.upsert")
         let testObject = ["name": "Bob", "age": 25] as [String : Any]
-        Backendless.shared.data.ofTable("TestClass").save(entity: testObject, responseHandler: { createdObject in
+        Backendless.shared.data.ofTable("TestClass").save(entity: testObject, isUpsert: false, responseHandler: { createdObject in
             var objectToUpdate = createdObject
             objectToUpdate["age"] = 30
             let uow = UnitOfWork()